+++ apps/backoffice-v2/src/pages/Entity/get-post-remove-decision-event-name.ts
@@ -1,0 +1,11 @@
+import { TWorkflowById } from '../../domains/workflows/fetchers';
+import { CommonWorkflowEvent } from '@ballerine/common';
+
+export function getPostRemoveDecisionEventName(workflow: TWorkflowById): string | undefined {
+  if (
+    !workflow?.workflowDefinition?.config?.workflowLevelResolution &&
+    workflow?.nextEvents?.includes(CommonWorkflowEvent.RETURN_TO_REVIEW)
+  ) {
+    return CommonWorkflowEvent.RETURN_TO_REVIEW;
+  }
+}