+++ services/workflows-service/scripts/workflows/kyc-email-process-example.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from '@prisma/client';
 import { env } from '../../src/env';
-import { StateTag } from '@ballerine/common';
+import { StateTag, WorkflowDefinitionVariant } from '@ballerine/common';
 
 export const kycEmailSessionDefinition = {
   id: 'kyc_email_session_example',
@@ -15,7 +15,7 @@ export const kycEmailSessionDefinition = {
       idle: {
         tags: [StateTag.PENDING_PROCESS],
         on: {
-          START: 'get_kyc_session',
+          start: 'get_kyc_session',
         },
       },
       get_kyc_session: {
@@ -164,6 +164,7 @@ export const kycEmailSessionDefinition = {
     },
   },
   isPublic: true,
+  variant: WorkflowDefinitionVariant.DEFAULT,
 };
 
 export const generateKycSessionDefinition = async (prismaClient: PrismaClient) => {